+++ frontend/dialob-composer-material/src/components/variables/ContextVariables.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { IconButton, TableBody, TableCell, TableHead, TableRow, Typography } from "@mui/material"
-import { ContextVariable, isContextVariable, useComposer } from "../../dialob"
+import { ContextVariable, useComposer } from "../../dialob"
 import { Add } from "@mui/icons-material";
 import { FormattedMessage } from "react-intl";
 import Tree, { TreeData, TreeDestinationPosition, TreeSourcePosition, moveItemOnTree } from "@atlaskit/tree";
@@ -8,6 +8,7 @@ import { INIT_TREE, buildTreeFromVariables } from "../../utils/TreeUtils";
 import { BorderedTable } from "../TableEditorComponents";
 import ContextVariableRow from "./ContextVariableRow";
 import { VariableProps } from "./VariableComponents";
+import { isContextVariable } from "../../utils/ItemUtils";
 
 const renderVariableItem = (props: VariableProps) => {
   const { item, provided, onClose } = props;
