+++ frontend/dialob-composer-material/src/intl/en.ts
@@ -9,10 +9,10 @@ const en = {
   'header.lists': 'Lists',
   'header.options': 'Options',
   'header.version': 'Version',
-  'header.help': 'Help',
+  'header.help': 'Dialob Composer User Guide',
   'header.search': 'Search',
   'header.search.hint': 'Enter search keyword',
-  'header.download': 'Download from as JSON',
+  'header.download': 'Download form as JSON',
   'header.status.fatal': 'Current form status: FATAL',
   'header.status.error': 'Current form status: ERROR',
   'header.status.warning': 'Current form status: WARNING',
