+++ frontend/dialob-composer-material/src/utils/SearchUtils.ts
