+++ frontend/dialob-composer-material/src/utils/ValidateUtils.ts
